front-back end integration for user setting page
